Final checking and cleanup on my end. Imo this should be ready for testing.
I'm not very good with C and this is my first code for Darkplaces. Therefore my coding may not be that great, so if anyone with more experience could check and optimize this I would be grateful. Please message me on IRC for anything regarding this patch (Taoki).

From: MirceaKitsune <sonichedgehog_hyperblast00@yahoo.com>

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@10134 d7cf8633-e32d-0410-b094-e92efae38249