finalize DP_CSQC_BINDMAPS: functions: 	string(float key[, float bindmap]) getkeybind = #342; 	float(float key, string bind[, float bindmap]) setkeybind = #630; 	vector(void) getbindmaps = #631; 	float(vector bm) setbindmaps(vector bm) = #632; 	string findkeysforcommand(string command[, float bindmap]) = #610; /* DEPRECATED csqc: #521 */ 	string keynumtostring(float keynum) = #340; /* menu: #609, DEPRECATED csqc: #520 */ 	float stringtokeynum(string key) = #341; /* DEPRECATED menu: #614 */
The bindmap argument is OPTIONAL, and maybe should be declared as ... in the definitons.
The old alternative builtin numbers are hereby deprecated; they come from esteel's unofficial extension to csqc to provide findkeysforcommand with a different builtin number than the menuqc one, and from menuqc builtins that now are part of the "more official" csqc spec in other builtin number, apart from #609 in menuqc which was documented as part of menuqc before.

From: Rudolf Polzer <divverent@alientrap.org>

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@10273 d7cf8633-e32d-0410-b094-e92efae38249
::stable-branch::merge=58ecffb6f73f96936bf8c21853570d7da483d0f3