call SetDllDirectory on Win32 too
Usage: put DLLs in a bin32/ subdir for that
If not done, binaries work as before.

This is the only way to have a 32bit and 64bit executable version in the same
directory! Otherwise Windows will abort DLL search at the 32bit DLL, which
fails to load into a 64bit process.

NOTE: this requires WinXP SP1 and above. To compile with this functionality
disabled, compile with -DDONT_USE_SETDLLDIRECTORY added to the CFLAGS.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@11168 d7cf8633-e32d-0410-b094-e92efae38249