fix interaction of blendfunc with rtlights
basically, we exclude a number of CUSTOMBLEND blendfuncs from receiving rtlights
the rationale is not that doing this is always correct - it is not in some
cases (e.g. for crazy stuff like GL_ONE GL_SRC_COLOR). The rationale is that
the current behaviour for these is wrong too.

All surfaces that now receive rtlights, receive it correctly.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@11326 d7cf8633-e32d-0410-b094-e92efae38249