Overhauled shadowmap rendering to use an atlas texture, this is overall a speed loss on default settings (-7.7% fps on timedemo demo1, or -3% with r_shadow_deferred) but a gain on realtime world lighting (+21.1% normal, +1.7% with r_shadow_deferred) on NVIDIA GTX 680.
On video drivers with higher framebuffer switching costs, this should be a framerate gain in all situations.

Feature-wise this brings support (in future) for shadowmapping on transparent lit surfaces, and imposes an upper bound on how many pixels of shadowmaps will be rendered in a frame (shadowmap resolution automatically becomes more coarse on frames where they do not fit).

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12281 d7cf8633-e32d-0410-b094-e92efae38249