First massive rename and move
A lot of these functions belong on their respective side of the engine.
host.c and host_cmd.c appear to be a dumping ground for misc functions
that Quakeworld engines have a proper place for. This is part of the
moving operation. There's still plenty more functions that need to be
put in their respective places.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12717 d7cf8633-e32d-0410-b094-e92efae38249