Eradicated the CVAR_NQUSERINFOHACK flag, for good
This is a massive commit, admittedly. But pretty much every single
change goes together and this could not be split up, unfortunately.

I implemented a few new functions and turned a lot of the userinfo keys
into cvars, just like their QW counterparts. CMD_INITWAIT is also gone
and replaced by CMD_USERINFO to denote special server-from-client
commands that specifically change a userinfo string of a given client.

I'm sure there's a cleaner way to do this, especially the holy callback
trinity I implemented to avoid an infinite loop when synchronizing
topcolor and bottomcolor with _cl_color and vice versa. But, if this
gets rid of decade old legacy cruft, and makes it all easier to maintain
down the road, that's a plus for me.

git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@12722 d7cf8633-e32d-0410-b094-e92efae38249