prvm: Initial experimental implementation of support for 32-bit progs offsets
This implementation is may be incomplete and also includes some fixes
to some opcodes. This shouldn't break anything in its current state.

Thanks to Spoike for the patch. He also left a few comments on some
compatibility issues DP ought to fix, which can be found by grepping
"spike FIXME".