cmake: Initial working implementation of cmake build system
This took way too long but I'm happy to be able to commit something
that at least builds the engine. It's a mess but I gotta start
somewhere. It is based on the build system of an old fork that I wrote
for "fun".

dpconfig.cmake is a special file you can drop into your project root
and it'll control the build configuration. I might look into a better
way of doing something like that.

I also added cmake-variants.json for vscode users and updated the
.gitignore appropriately.