NudgeOutOfSolid: fix "random" failures on slightly inclined planes
History:
This epsilon value was first added to the nudge distance in
e3f5fe5206ac395b5c789a4151cb36c339257386
and was replaced with bbox expansion in
2d472cd3f605bf9b8e6823105eb8d060cf35a1bc

With the expansion approach we need to compare to separation to make
"good location" detection reliable, partly to avoid float precision
problems, partly to avoid compatibility problems with normal TraceBox.
For example: we drop an entity to floor successfully with TraceBox, but
then NudgeOutOfSolid decides it's in solid.

Also updated the startsolid conditions to suit the 2-pass approach.

Signed-off-by: bones_was_here <bones_was_here@xonotic.au>