Fix OP_???_IF math.
Previously 5 / 1.25 returned 5 due to rounding the 1.25 to int and
performing an integer division.

Now 5 / 1.25 will return 4 due to converting the 5 to floating point and
performing a floating division.

Same applied to +, - and * for int and float types in this order.

Which is a lot more what one would expect based on what C does, and
also matches FTEQW.