README: Reword disclaimer
Due to regrettable circumstances, core developers of the engine are no longer a part of Team Xonotic. Therefore, the Xonotic GitLab mirror is no longer maintained by DarkPlaces maintainers.

The intention of the wording was to clarify a possible misconception that Xonotic == DarkPlaces, arising from that fact, and the fact that we previously used some of Xonotic's infrastructure.

The previous wording may come off as unduly hostile on further consideration and this is personally not my intention.